+++ curator/cli.py
@@ -32,6 +32,7 @@ CLASS_MAP = {
     'forcemerge' : ForceMerge,
     'open' : Open,
     'replicas' : Replicas,
+    'restore' : Restore,
     'snapshot' : Snapshot,
 }
 
@@ -52,17 +53,23 @@ def process_action(client, config, **kwargs):
     mykwargs = {}
 
     if action in CLASS_MAP:
-        if action == 'delete_indices':
-            mykwargs['master_timeout'] = (
-                kwargs['master_timeout'] if 'master_timeout' in kwargs else 30)
         # deepcopy guarantees clean copies of the defaults, and nothing getting
         # altered in "pass by reference," which was happening in testing.
         mykwargs = copy.deepcopy(ACTION_DEFAULTS[action])
-        logger.debug('MYKWARGS = {0}'.format(mykwargs))
         action_class = CLASS_MAP[action]
     else:
         raise ConfigurationError(
             'Unrecognized action: {0}'.format(action))
+
+    # Override some settings...
+    if action == 'delete_indices':
+        mykwargs['master_timeout'] = (
+            kwargs['master_timeout'] if 'master_timeout' in kwargs else 30)
+    if action == 'allocation' or action == 'replicas':
+        # Setting the operation timeout to the client timeout
+        mykwargs['timeout'] = (
+            kwargs['timeout'] if 'timeout' in kwargs else 30)
+    logger.debug('MYKWARGS = {0}'.format(mykwargs))
 
     ### Update the defaults with whatever came with opts, minus any Nones
     mykwargs.update(prune_nones(opts))
@@ -88,8 +95,8 @@ def process_action(client, config, **kwargs):
             action_obj.remove(removes)
     elif action == 'create_index':
         action_obj = action_class(client, **mykwargs)
-    elif action == 'delete_snapshots':
-        logger.debug('Running "delete_snapshots"')
+    elif action == 'delete_snapshots' or action == 'restore':
+        logger.debug('Running "{0}"'.format(action))
         slo = SnapshotList(client, repository=opts['repository'])
         slo.iterate_filters(config)
         # We don't need to send this value to the action
@@ -153,8 +160,9 @@ def cli(config, dry_run, action_file):
         sys.exit(1)
     test_client_options(client_args)
 
-    # Create a client object
-    client = get_client(**client_args)
+    # Extract this and save it for later, in case there's no timeout_override.
+    default_timeout = client_args.pop('timeout')
+    logger.debug('default_timeout = {0}'.format(default_timeout))
     #########################################
     ### Start working on the actions here ###
     #########################################
@@ -175,12 +183,10 @@ def cli(config, dry_run, action_file):
         action_disabled = actions[idx]['options'].pop('disable_action', False)
         continue_if_exception = (
             actions[idx]['options'].pop('continue_if_exception', False))
+        timeout_override = actions[idx]['options'].pop('timeout_override', None)
         logger.debug(
             'continue_if_exception = {0}'.format(continue_if_exception))
-        kwargs = {}
-        kwargs['master_timeout'] = (
-            client_args['timeout'] if client_args['timeout'] <= 300 else 300)
-        kwargs['dry_run'] = dry_run
+        logger.debug('timeout_override = {0}'.format(timeout_override))
 
         ### Skip to next action if 'disabled'
         if action_disabled:
@@ -189,6 +195,23 @@ def cli(config, dry_run, action_file):
                 'True'.format(action)
             )
             continue
+
+        # Override the timeout, if specified, otherwise use the default.
+        if type(timeout_override) == type(int()):
+            client_args['timeout'] = timeout_override
+        else:
+            client_args['timeout'] = default_timeout
+
+        # Set up action kwargs
+        kwargs = {}
+        kwargs['master_timeout'] = (
+            client_args['timeout'] if client_args['timeout'] <= 300 else 300)
+        kwargs['dry_run'] = dry_run
+        kwargs['timeout'] = client_args['timeout']
+
+        # Create a client object for each action...
+        client = get_client(**client_args)
+        logger.debug('client is {0}'.format(type(client)))
         ##########################
         ### Process the action ###
         ##########################