+++ test/unit/test_action_replicas.py
@@ -45,6 +45,16 @@ class TestActionReplicas(TestCase):
         ilo = curator.IndexList(client)
         ro = curator.Replicas(ilo, count=0)
         self.assertIsNone(ro.do_action())
+    def test_do_action_wait(self):
+        client = Mock()
+        client.indices.get_settings.return_value = testvars.settings_one
+        client.cluster.state.return_value = testvars.clu_state_one
+        client.indices.stats.return_value = testvars.stats_one
+        client.indices.put_settings.return_value = None
+        client.cluster.health.return_value = None
+        ilo = curator.IndexList(client)
+        ro = curator.Replicas(ilo, count=1, wait_for_completion=True)
+        self.assertIsNone(ro.do_action())
     def test_do_action_raises_exception(self):
         client = Mock()
         client.indices.get_settings.return_value = testvars.settings_one